Moved firebase config to its own file. Initialise Vue after Firebase initialises on reload
